Factor out setting of web server dir
